backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
python-dotenv
djangorestframework-simplejwt
twilio
google-generativeai
pymongo
google-genai
gunicorn
pyzbar
opencv-python
redis
django-redis